Fixed type error in dlr_c.h

+++ src/dlr_c/dlr_c.h
@@ -18,7 +18,7 @@ void c_dlr_it2cf_init(int *rank, double *dlrrf, double *dlrit, double *it2cf, in
 
 void c_dlr_it2cf(int *r, double *it2cf, int *it2cfp, double *g, double *gc);
 
-void c_dlr_it2itr_init(int *r, double *dlrrf, double *dlrit, double *it2cf, double *it2cfp, double *it2itr);
+void c_dlr_it2itr_init(int *r, double *dlrrf, double *dlrit, double *it2cf, int *it2cfp, double *it2itr);
 
 void c_dlr_it2itr(int *r, double *it2itr, double *g, double *gr);
 
@@ -34,7 +34,7 @@ void c_dlr_convtens(double *beta, int *xi, int *r, double *dlrrf, double *dlrit,
 
 void c_dlr_convmat(int *r, double *it2cf, int *it2cfp, double *phi, double *g, double *gmat);
 
-void c_dlr_ipmat(double *beta, int *r, double *dlrit, double *dlrrf, double *it2cf, double *it2cfp, double *ipmat);
+void c_dlr_ipmat(double *beta, int *r, double *dlrit, double *dlrrf, double *it2cf, int *it2cfp, double *ipmat);
 
 void c_eqpts_rel(int *n, double *t);
